The Raspberry Pi build will now use the OpenGL 2.1 backend for ImGui.  Not exactly speedy, the Roller Coaster Tycoon demo seems to get about 1 frame per second.

+++ source/io/fszipnode.cpp
@@ -20,7 +20,7 @@ bool FsZipNode::moveToFileSystem(BoxedPtr<FsNode> node) {
         U8 buffer[4096];
         U32 read = from->readNative(buffer, 4096);
         while (read) {
-            if (::write(to, buffer, read) != read)
+            if (::write(to, buffer, read) != (int)read)
                 return false;
             read = from->readNative(buffer, 4096);
         }
